backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
python-dotenv
httpx
python-multipart
boto3
PyPDF2
python-docx
psycopg2-binary
numpy